cleaned up javascript folder
